crearAdmin working

+++ Backend/src/main/java/com/example/Analitica/repository/AdminRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AdminRepository extends CrudRepository<Admin, Integer> {
+public interface AdminRepository extends CrudRepository<Admin, Long> {
 
 }
